Consolidate test IFeatures implementations into shared TestFeatures class
