+++ tests/Aspire.Cli.Tests/DotNetSdkInstallerTests.cs
@@ -3,7 +3,6 @@
 
 using Microsoft.AspNetCore.InternalTesting;
 using System.Globalization;
-using Aspire.Cli.Configuration;
 using Aspire.Cli.DotNet;
 using Aspire.Cli.Resources;
 using Microsoft.Extensions.Configuration;
@@ -248,22 +247,4 @@ public class DotNetSdkInstallerTests
             })
             .Build();
     }
-}
-
-public class TestFeatures : IFeatures
-{
-    private readonly Dictionary<string, bool> _features = new();
-
-    public TestFeatures SetFeature(string featureName, bool value)
-    {
-        _features[featureName] = value;
-        return this;
-    }
-
-    public bool IsFeatureEnabled(string featureName, bool defaultValue = false)
-    {
-        return _features.TryGetValue(featureName, out var value) ? value : defaultValue;
-    }
-
-    public void LogFeatureState() { }
 }
